# Resample : Optimize inseparable case with no scaling