@johnhaddon 2023-11-20
Rebased to add a Changes.md entry, but otherwise unchanged. Merging.